Did requested changes
