+++ src/main/java/io/github/addoncommunity/galactifun/base/BaseRegistry.java
@@ -7,7 +7,7 @@ import io.github.addoncommunity.galactifun.api.universe.StarSystem;
 import io.github.addoncommunity.galactifun.base.milkyway.solarsystem.Earth;
 import io.github.addoncommunity.galactifun.base.milkyway.solarsystem.Mars;
 import io.github.addoncommunity.galactifun.base.milkyway.solarsystem.TheMoon;
-import io.github.addoncommunity.galactifun.base.milkyway.solarsystem.Venus;
+import io.github.addoncommunity.galactifun.base.milkyway.solarsystem.venus.Venus;
 import lombok.experimental.UtilityClass;
 
 /**
